arm64: dts: rcar: s4sk: add support of remote proc with openamp
* Use fixed clk for MFIS module
MFIS block module does have a clock but it is a no modifiable one.
MFIS_clk is fixed and always ON.

* Correct MFIS interrupts
MFIS channels have an IRQ for each. They are not the same as GEN3 and the ones
used by CR52 cores. Need to use the one that are coming from AR core to AP core.
 * MFISAREICR0 generate IRQ number 320
 * MFISAREICR1 generate IRQ number 322

Signed-off-by: Pierre Marzin <pierre.marzin@iot.bzh>